Refactored with numpy

+++ 2021/python/Inputs.py
@@ -10,7 +10,7 @@ def Day02():
 
 def Day03():
     import numpy as np
-    return np.genfromtxt("txtfiles/day03.txt", delimiter=1, dtype=np.uint16)
+    return np.genfromtxt("txtfiles/day03.txt", delimiter=1, dtype=np.uint8)
 
 def Day04():
     import numpy as np
